lab3/requirements.txt
# Core data processing libraries
pandas>=2.0.0
numpy>=1.24.0

# Machine learning and similarity calculations
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0

# HTTP requests for IMDb API
requests>=2.31.0

# Image processing for movie posters
Pillow>=10.0.0

